src/main/java/com/nbicocchi/exercises/oop/shapev2/Circle.java
package com.nbicocchi.exercises.oop.shapev2;

import java.awt.*;

public class Circle implements Shape {
    Color color;
    boolean filled;
    Point center;
    double radius;

    public Circle(Color color, boolean filled, Point center, double radius) {
        this.color = color;
        this.filled = filled;
        this.center = center;
        this.radius = radius;
    }

    public Color getColor() {
        return color;
    }

    public void setColor(Color color) {
        this.color = color;
    }

    public boolean isFilled() {
        return filled;
    }

    public void setFilled(boolean filled) {
        this.filled = filled;
    }

    public Point getCenter() {
        return center;
    }

    public void setCenter(Point center) {
        this.center = center;
    }

    public double getRadius() {
        return radius;
    }

    public void setRadius(double radius) {
        this.radius = radius;
    }

    @Override
    public double getPerimeter() {
        return 2 * Math.PI * radius;
    }

    @Override
    public double getArea() {
        return Math.pow(radius, 2) * Math.PI;
    }

    @Override
    public void move(Point movement) {
        center.translate(movement.x, movement.y);
    }

    @Override
    public void resize(double scale) {
        radius *= scale;
    }

    @Override
    public String toString() {
        return "Circle{" + "center=" + center + ", radius=" + radius + '}';
    }
}
